Make user fragment abstract and add button to view other users profile
